[chore] 리펙토리
