+++ app/src/main/java/com/konkuk/moru/core/component/routinedetail/RoutineStepEditableList.kt
@@ -1,8 +1,17 @@
-package com.konkuk.moru.presentation.myroutines.component
+package com.konkuk.moru.core.component.routinedetail
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.gestures.detectDragGesturesAfterLongPress
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListScope
 import androidx.compose.foundation.lazy.LazyListState
@@ -11,13 +20,15 @@ import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material3.*
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.input.pointer.consumeAllChanges
 import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
@@ -25,6 +36,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.zIndex
 import com.konkuk.moru.data.model.RoutineStep
 import com.konkuk.moru.data.model.RoutineStepActions
+import com.konkuk.moru.presentation.myroutines.component.findNewIndex
 import com.konkuk.moru.ui.theme.MORUTheme
 
 /**
@@ -60,8 +72,8 @@ fun LazyListScope.routineStepEditableList(
 
         val dragHandleModifier = if (isEditMode) {
             Modifier.pointerInput(Unit) {
-                var localOffset = 0f          // 누적 오프셋
-                val startIndex = index        // ← 드래그 시작 인덱스를 로컬에 캡처!
+                var localOffset = 0f
+                val startIndex = index
 
                 detectDragGesturesAfterLongPress(
                     onDragStart = {
