+++ app/src/main/java/com/konkuk/moru/data/model/AppInfo.kt
@@ -1,0 +1,6 @@
+package com.konkuk.moru.data.model
+
+data class AppInfo(
+    val name: String,
+    val iconUrl: String?
+)
