+++ app/src/main/java/com/konkuk/moru/data/model/Routine.kt
@@ -41,26 +41,5 @@ data class Routine(
     val usedApps: List<AppInfo> = emptyList()
 )
 
-data class AppInfo(
-    val name: String,
-    val iconUrl: String?
-)
 
-/**
- * 루틴의 각 단계를 나타내는 클래스
- */
-data class RoutineStep(
-    val id: UUID = UUID.randomUUID(),
-    val name: String,
-    val duration: String // "00:00" 형식
-)
 
-/**
- * 비슷한 루틴 정보를 나타내는 클래스
- */
-data class SimilarRoutine(
-    val id: Int,
-    val imageUrl: String?,
-    val name: String,
-    val tag: String
-)
