+++ app/src/main/java/com/konkuk/moru/data/model/SimilarRoutine.kt
@@ -1,0 +1,8 @@
+package com.konkuk.moru.data.model
+
+data class SimilarRoutine(
+    val id: Int,
+    val imageUrl: String?,
+    val name: String,
+    val tag: String
+)
