+++ app/src/main/java/com/konkuk/moru/data/model/TagDto.kt
@@ -1,0 +1,7 @@
+package com.konkuk.moru.data.model
+
+data class TagDto(
+    val id: Int,
+    val name: String,
+    val isSelected: Boolean = false
+)
