+++ app/src/main/java/com/konkuk/moru/presentation/routinefeed/component/search/HashTagSearchField.kt
@@ -1,0 +1,72 @@
+package com.konkuk.moru.presentation.routinefeed.component.search
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.BasicTextField
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.input.ImeAction
+import androidx.compose.ui.unit.dp
+import com.konkuk.moru.R
+import com.konkuk.moru.ui.theme.MORUTheme
+
+
+@Composable
+internal fun HashTagSearchField(
+    value: String,
+    onValueChange: (String) -> Unit,
+    onSearch: () -> Unit = {}
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(45.dp)
+            .clip(RoundedCornerShape(24.dp))
+            .border(1.dp, MORUTheme.colors.lightGray, shape = RoundedCornerShape(24.dp))
+            .background(Color(0xFFF5F6F8))
+            .padding(horizontal = 16.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.ic_magnifying),
+            contentDescription = "Search Icon",
+            modifier = Modifier.size(20.dp)
+        )
+        Spacer(modifier = Modifier.width(8.dp))
+        BasicTextField(
+            value = value,
+            onValueChange = onValueChange,
+            singleLine = true,
+            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Search),
+            keyboardActions = KeyboardActions(onSearch = { onSearch() }),
+            textStyle = MORUTheme.typography.time_R_14.copy(color = Color.Black),
+            modifier = Modifier.weight(1f),
+            decorationBox = { innerTextField ->
+                if (value.isEmpty()) {
+                    Text(
+                        text = "태그를 검색해 보세요.",
+                        style = MORUTheme.typography.time_R_14,
+                        color = MORUTheme.colors.mediumGray
+                    )
+                }
+                innerTextField()
+            }
+        )
+    }
+}
