+++ app/src/main/java/com/konkuk/moru/presentation/routinefeed/component/search/SearchComponents.kt
@@ -49,14 +49,11 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.konkuk.moru.R
 import com.konkuk.moru.core.component.chip.MoruChip
+import com.konkuk.moru.presentation.myactivity.screen.TagDto
 import com.konkuk.moru.ui.theme.MORUTheme
 
 // --- Data Transfer Object for Tags ---
-data class TagDto(
-    val id: Int,
-    val name: String,
-    val isSelected: Boolean = false
-)
+
 
 fun generateDummyTags(): List<TagDto> {
     val sampleTagWords = listOf(
@@ -70,198 +67,6 @@ fun generateDummyTags(): List<TagDto> {
     return tags.map { tag -> selected.find { it.id == tag.id }?.copy(isSelected = true) ?: tag }
 }
 
-// --- Common UI Components ---
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-internal fun RoutineSearchTopAppBar(
-    query: String,
-    onQueryChange: (String) -> Unit,
-    onSearch: () -> Unit,
-    onNavigateBack: () -> Unit,
-    placeholderText: String
-) {
-    TopAppBar(
-        colors = TopAppBarDefaults.topAppBarColors(containerColor = Color.White),
-        title = {
-            BasicTextField(
-                value = query,
-                onValueChange = onQueryChange,
-                modifier = Modifier.fillMaxWidth(),
-                textStyle = MORUTheme.typography.body_SB_16.copy(color = MORUTheme.colors.black),
-                singleLine = true,
-                cursorBrush = SolidColor(MORUTheme.colors.black),
-                keyboardOptions = KeyboardOptions(imeAction = ImeAction.Search),
-                keyboardActions = KeyboardActions(onSearch = { onSearch() }),
-                decorationBox = { innerTextField ->
-                    if (query.isEmpty()) {
-                        Text(
-                            text = placeholderText,
-                            style = MORUTheme.typography.desc_M_16,
-                            color = MORUTheme.colors.lightGray
-                        )
-                    }
-                    innerTextField()
-                }
-            )
-        },
-        navigationIcon = {
-            IconButton(onClick = onNavigateBack) {
-                Icon(
-                    Icons.AutoMirrored.Filled.KeyboardArrowLeft,
-                    contentDescription = "뒤로가기",
-                    modifier = Modifier.size(32.dp),
-                    tint = MORUTheme.colors.black
-                )
-            }
-        },
-        actions = {
-            IconButton(onClick = onSearch) {
-                Icon(
-                    painter = painterResource(id = R.drawable.ic_magnifying),
-                    contentDescription = "검색",
-                    modifier = Modifier.size(24.dp),
-                    tint = MORUTheme.colors.black
-                )
-            }
-        }
-    )
-}
-
-@Composable
-internal fun TagSearchBar(text: String, onClick: () -> Unit) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(45.dp)
-            .clip(RoundedCornerShape(100.dp))
-            .background(Color(0xFFF5F6F8))
-            .border(1.dp, MORUTheme.colors.lightGray, RoundedCornerShape(100.dp))
-            .clickable { onClick() }
-            .padding(horizontal = 16.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Icon(
-            painter = painterResource(id = R.drawable.ic_magnifying),
-            contentDescription = "태그 검색",
-            tint = MORUTheme.colors.darkGray,
-            modifier = Modifier.size(20.dp)
-        )
-        Spacer(modifier = Modifier.size(8.dp))
-        Text(
-            text = text,
-            style = MORUTheme.typography.time_R_14,
-            color = MORUTheme.colors.mediumGray
-        )
-    }
-}
-
-@Composable
-internal fun RecentSearchesHeader(onDeleteAll: () -> Unit) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        Text(text = "최근 검색어", style = MORUTheme.typography.body_SB_14)
-        Text(
-            text = "전체 삭제",
-            style = MORUTheme.typography.desc_M_12,
-            color = MORUTheme.colors.mediumGray,
-            modifier = Modifier.clickable { onDeleteAll() }
-        )
-    }
-}
-
-@Composable
-internal fun SortButtons(selectedOption: String, onOptionSelected: (String) -> Unit) {
-    Row(
-        modifier = Modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.End,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        SortButton("최신순", selectedOption == "최신순") { onOptionSelected("최신순") }
-        Spacer(modifier = Modifier.size(8.dp))
-        SortButton("인기순", selectedOption == "인기순") { onOptionSelected("인기순") }
-    }
-}
-
-@Composable
-private fun SortButton(text: String, isSelected: Boolean, onClick: () -> Unit) {
-    val backgroundColor = if (isSelected) Color.Black else MORUTheme.colors.veryLightGray
-    val textColor = if (isSelected) Color.White else MORUTheme.colors.mediumGray
-    Box(
-        modifier = Modifier
-            .clip(RoundedCornerShape(100.dp))
-            .background(backgroundColor)
-            .clickable { onClick() }
-            .padding(horizontal = 12.dp, vertical = 6.dp)
-    ) {
-        Text(text, color = textColor, style = MORUTheme.typography.body_SB_14)
-    }
-}
-
-@Composable
-internal fun BackTitle(title: String, onNavigateBack: () -> Unit) {
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        IconButton(onClick = onNavigateBack) {
-            Icon(
-                Icons.AutoMirrored.Filled.KeyboardArrowLeft,
-                contentDescription = "Back",
-                modifier = Modifier.size(32.dp)
-            )
-        }
-        Spacer(modifier = Modifier.width(8.dp))
-        Text(text = title, style = MORUTheme.typography.title_B_20)
-    }
-}
-
-@Composable
-internal fun HashTagSearchField(
-    value: String,
-    onValueChange: (String) -> Unit,
-    onSearch: () -> Unit = {}
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(45.dp)
-            .clip(RoundedCornerShape(24.dp))
-            .border(1.dp, MORUTheme.colors.lightGray, shape = RoundedCornerShape(24.dp))
-            .background(Color(0xFFF5F6F8))
-            .padding(horizontal = 16.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Image(
-            painter = painterResource(id = R.drawable.ic_magnifying),
-            contentDescription = "Search Icon",
-            modifier = Modifier.size(20.dp)
-        )
-        Spacer(modifier = Modifier.width(8.dp))
-        BasicTextField(
-            value = value,
-            onValueChange = onValueChange,
-            singleLine = true,
-            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Search),
-            keyboardActions = KeyboardActions(onSearch = { onSearch() }),
-            textStyle = MORUTheme.typography.time_R_14.copy(color = Color.Black),
-            modifier = Modifier.weight(1f),
-            decorationBox = { innerTextField ->
-                if (value.isEmpty()) {
-                    Text(
-                        text = "태그를 검색해 보세요.",
-                        style = MORUTheme.typography.time_R_14,
-                        color = MORUTheme.colors.mediumGray
-                    )
-                }
-                innerTextField()
-            }
-        )
-    }
-}
 
 @Composable
 internal fun TagSectionHeader(title: String) {
@@ -338,87 +143,6 @@ internal fun SelectedTagChip(
     )
 }
 
-@Composable
-internal fun TagInputSearchBar(
-    modifier: Modifier = Modifier,
-    selectedTags: List<String>,
-    onDeleteTag: (String) -> Unit,
-    onAddTag: (String) -> Unit,
-    onNavigateToTagSearch: () -> Unit
-) {
-    var tagInput by remember { mutableStateOf("") }
-    val focusManager = LocalFocusManager.current
-
-    fun submitTag() {
-        if (tagInput.isNotBlank()) {
-            val tag = if (tagInput.startsWith("#")) tagInput else "#$tagInput"
-            onAddTag(tag)
-            tagInput = ""
-            focusManager.clearFocus()
-        }
-    }
-
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .height(IntrinsicSize.Min)
-            .clip(RoundedCornerShape(100.dp))
-            .background(MORUTheme.colors.veryLightGray)
-            .border(1.dp, MORUTheme.colors.lightGray, RoundedCornerShape(100.dp))
-            .padding(start = 12.dp, end = 4.dp, top = 4.dp, bottom = 4.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-
-        /* ---------- ① 태그 + 입력창 : 남는 공간 전부 ---------- */
-        LazyRow(
-            modifier = Modifier
-                .weight(1f)              // ★ 핵심: 아이콘을 제외한 모든 공간 차지
-                .padding(end = 8.dp),    // 아이콘과 살짝 간격
-            horizontalArrangement = Arrangement.spacedBy(8.dp),
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            /* 선택된 태그 칩 */
-            items(selectedTags) { tag ->
-                SelectedTagChip(text = tag) { onDeleteTag(tag) }
-            }
-
-            /* 입력창 */
-            item {
-                BasicTextField(
-                    value = tagInput,
-                    onValueChange = { tagInput = it },
-                    singleLine = true,
-                    textStyle = MORUTheme.typography.time_R_14
-                        .copy(color = MORUTheme.colors.black),
-                    keyboardOptions = KeyboardOptions(imeAction = ImeAction.Search),
-                    keyboardActions = KeyboardActions { submitTag() },
-                    decorationBox = { inner ->
-                        if (tagInput.isEmpty() && selectedTags.isEmpty()) {
-                            Text(
-                                text = "#태그 검색",
-                                style = MORUTheme.typography.time_R_14,
-                                color = MORUTheme.colors.mediumGray
-                            )
-                        }
-                        inner()
-                    }
-                )
-            }
-        }
-
-        /* ---------- ② “+” 아이콘 : 항상 오른쪽 ---------- */
-        IconButton(
-            onClick = onNavigateToTagSearch,
-            modifier = Modifier.size(32.dp)
-        ) {
-            Icon(
-                imageVector = Icons.Default.Add,
-                contentDescription = "태그 추가",
-                tint = MORUTheme.colors.mediumGray
-            )
-        }
-    }
-}
 
 
 @Preview
