+++ app/src/main/java/com/konkuk/moru/presentation/routinefeed/component/search/TagSearchBar.kt
@@ -1,0 +1,52 @@
+package com.konkuk.moru.presentation.routinefeed.component.search
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Icon
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import com.konkuk.moru.R
+import com.konkuk.moru.ui.theme.MORUTheme
+
+
+@Composable
+internal fun TagSearchBar(text: String, onClick: () -> Unit) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(45.dp)
+            .clip(RoundedCornerShape(100.dp))
+            .background(Color(0xFFF5F6F8))
+            .border(1.dp, MORUTheme.colors.lightGray, RoundedCornerShape(100.dp))
+            .clickable { onClick() }
+            .padding(horizontal = 16.dp),
+        verticalAlignment = Alignment.CenterVertically
+    ) {
+        Icon(
+            painter = painterResource(id = R.drawable.ic_magnifying),
+            contentDescription = "태그 검색",
+            tint = MORUTheme.colors.darkGray,
+            modifier = Modifier.size(20.dp)
+        )
+        Spacer(modifier = Modifier.size(8.dp))
+        Text(
+            text = text,
+            style = MORUTheme.typography.time_R_14,
+            color = MORUTheme.colors.mediumGray
+        )
+    }
+}
