+++ app/src/main/java/com/konkuk/moru/presentation/routinefeed/screen/main/RoutineFeedScreen.kt
@@ -43,7 +43,6 @@ fun RoutineFeedScreen(
     uiState: RoutineFeedUiState,
     onNotificationClick: () -> Unit,
 ) {
-    var searchQuery by remember { mutableStateOf("") }
     val liveUsers = DummyData.dummyLiveUsers
 
     val routineSections = remember {
